refactor: add old structure to new zap structure for payments
