+++ core/src/app/global-error.tsx
@@ -3,7 +3,7 @@ import "client-only";
 import "./globals.css";
 
 import { geist } from "@/fonts";
-import { ZapButton } from "@/zap/components/button";
+import { ZapButton } from "@/zap/components";
 import { BaseError } from "@/zap/lib/api/errors";
 
 interface ErrorBoundaryProps {
