+++ core/src/zap/auth/components/forms/forgot-password-form.tsx
@@ -18,7 +18,7 @@ import { Input } from "@/components/ui/input";
 import { useCooldown } from "@/hooks/utils/use-cooldown";
 import { SETTINGS } from "@/lib/settings";
 import { betterAuthClient } from "@/zap/auth/providers/better-auth/client";
-import { ZapButton } from "@/zap/components/button";
+import { ZapButton } from "@/zap/components";
 import { handleClientError } from "@/zap/errors/client";
 
 const formSchema = z.object({
