+++ core/src/zap/auth/components/social-provider-button.tsx
@@ -6,7 +6,7 @@ import { toast } from "sonner";
 import { Provider, ZAP_DEFAULT_SETTINGS } from "@/zap.config";
 import { AUTH_ICONS } from "@/zap/auth/components";
 import { betterAuthClient } from "@/zap/auth/providers/better-auth/client";
-import { ZapButton } from "@/zap/components/button";
+import { ZapButton } from "@/zap/components";
 import { AuthenticationError } from "@/zap/errors";
 import { handleClientError } from "@/zap/errors/client";
