+++ core/src/zap/components/index.ts
@@ -1,0 +1,1 @@
+export * from "@/zap/components/button";
