+++ core/src/zap/payments/components/index.ts
@@ -1,0 +1,4 @@
+export * from "@/zap/payments/components/billing-cards";
+export * from "@/zap/payments/components/faq";
+export * from "@/zap/payments/components/price-display";
+export * from "@/zap/payments/components/pricing-toggle";
