+++ core/src/zap/payments/data/index.ts
@@ -1,0 +1,22 @@
+export const BILLING_FAQ = [
+  {
+    question: "Can I cancel anytime?",
+    answer:
+      "Yes, you can cancel your subscription at any time. Your access will continue until the end of your billing period.",
+  },
+  {
+    question: "Do you offer refunds?",
+    answer:
+      "We offer a 30-day money-back guarantee. If you're not satisfied with your Pro subscription, contact our support team for a full refund.",
+  },
+  {
+    question: "What payment methods do you accept?",
+    answer:
+      "We accept all major credit cards, PayPal, and other payment methods through our secure payment processor.",
+  },
+  {
+    question: "Can I upgrade or downgrade my plan?",
+    answer:
+      "Yes, you can switch between monthly and yearly plans at any time. Changes will be reflected in your next billing cycle.",
+  },
+];
