+++ core/src/zap/payments/providers/polar/utils.ts
@@ -1,7 +1,5 @@
 import type { CheckoutOptions } from "@polar-sh/better-auth";
 
-import { type ProductMetadata, PRODUCTS_METADATA } from "@/zap.config";
-
 type ExtractProductType<T> = T extends {
   products?: infer P | (() => Promise<infer Q>);
 }
@@ -14,18 +12,3 @@ type ExtractProductType<T> = T extends {
         : P
   : never;
 export type Product = ExtractProductType<CheckoutOptions>;
-
-export function getProducts() {
-  const products = Object.values(PRODUCTS_METADATA);
-  return products;
-}
-
-export function getProduct({
-  products,
-  productId,
-}: {
-  products: ProductMetadata[];
-  productId: string;
-}) {
-  return products.find((product) => product.productId === productId) || null;
-}
