+++ core/src/zap/payments/utils.ts
@@ -1,4 +1,23 @@
-import type { ProductMetadata, RecurringInterval } from "@/zap.config";
+import {
+  type ProductMetadata,
+  PRODUCTS_METADATA,
+  type RecurringInterval,
+} from "@/zap.config";
+
+export function getProducts() {
+  const products = Object.values(PRODUCTS_METADATA);
+  return products;
+}
+
+export function getProduct({
+  products,
+  productId,
+}: {
+  products: ProductMetadata[];
+  productId: string;
+}) {
+  return products.find((product) => product.productId === productId) || null;
+}
 
 export function getBillingDetails(product: ProductMetadata, isYearly: boolean) {
   const billingKey = isYearly ? "yearly" : "monthly";
