+++ core/src/zap/waitlist/components/waitlist-form.tsx
@@ -11,7 +11,7 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { cn } from "@/lib/utils";
-import { ZapButton } from "@/zap/components/button";
+import { ZapButton } from "@/zap/components";
 import { useWaitlist } from "@/zap/hooks/waitlist/use-waitlist";
 
 export function WaitlistForm() {
